requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
pillow==10.4.0
setuptools==75.1.0